+++ lib/core/failures/feature/auth/auth_failure.dart
@@ -2,9 +2,6 @@ import '../../app_failure.dart';
 
 /// Auth application error.
 sealed class AuthFailure extends AppFailure {
-  /// Error code.
-  abstract final String code;
-
   /// Creates an instance of [AuthFailure].
   const AuthFailure(
     super.message, {
@@ -13,19 +10,70 @@ sealed class AuthFailure extends AppFailure {
   });
 }
 
-/// Unknown authentication error.
-final class UnknownAuthFailure extends AuthFailure {
-  @override
-  final String code;
-
-  /// Original error message.
-  final String? originalMessage;
-
-  /// Creates an instance of [UnknownAuthFailure].
-  const UnknownAuthFailure(
-    this.code, {
-    this.originalMessage,
+/// Invalid credentials error.
+final class InvalidCredentialsFailure extends AuthFailure {
+  /// Creates an instance of [InvalidCredentialsFailure].
+  const InvalidCredentialsFailure({
     super.parentException,
     super.stackTrace,
-  }) : super(originalMessage ?? 'Unknown authentication error');
+  }) : super('Неверные учетные данные');
+}
+
+/// Validation failed error.
+final class ValidationFailedFailure extends AuthFailure {
+  /// Field-specific validation errors.
+  final Map<String, List<String>> fieldErrors;
+
+  /// Creates an instance of [ValidationFailedFailure].
+  const ValidationFailedFailure({
+    super.parentException,
+    super.stackTrace,
+    this.fieldErrors = const {},
+  }) : super('Проверьте введенные данные и попробуйте снова');
+}
+
+/// Email already verified error.
+final class EmailAlreadyVerifiedFailure extends AuthFailure {
+  /// Creates an instance of [EmailAlreadyVerifiedFailure].
+  const EmailAlreadyVerifiedFailure({
+    super.parentException,
+    super.stackTrace,
+  }) : super('Email уже подтвержден. Войдите в аккаунт');
+}
+
+/// Email not verified error.
+final class EmailNotVerifiedFailure extends AuthFailure {
+  /// Creates an instance of [EmailNotVerifiedFailure].
+  const EmailNotVerifiedFailure({
+    super.parentException,
+    super.stackTrace,
+  }) : super('Подтвердите email, чтобы продолжить');
+}
+
+/// Auth Rate limited error.
+final class AuthRateLimitedFailure extends AuthFailure {
+  /// Creates an instance of [AuthRateLimitedFailure].
+  const AuthRateLimitedFailure({
+    super.parentException,
+    super.stackTrace,
+  }) : super('Слишком много попыток. Попробуйте позже');
+}
+
+/// Unauthorized error.
+final class UnauthorizedAuthFailure extends AuthFailure {
+  /// Creates an instance of [UnauthorizedAuthFailure].
+  const UnauthorizedAuthFailure({
+    super.parentException,
+    super.stackTrace,
+    // logout/redirect without showing a snackbar
+  }) : super('');
+}
+
+/// Unknown authentication error.
+final class UnknownAuthFailure extends AuthFailure {
+  /// Creates an instance of [UnknownAuthFailure].
+  const UnknownAuthFailure({
+    super.parentException,
+    super.stackTrace,
+  }) : super('Не удалось выполнить действие. Попробуйте снова');
 }