+++ lib/core/network/api_paths.dart
@@ -1,7 +1,10 @@
-import '../../core/env/env.dart';
+import '../env/env.dart';
 
 /// API paths used in the application.
 abstract class ApiPaths {
   /// The base URL for the API.
   static final String baseUrl = Env.apiUrl;
+
+  /// The endpoint for refreshing the access token.
+  static const String refresh = '/refresh';
 }