+++ lib/core/services/token_storage/token_storage.dart
@@ -1,0 +1,11 @@
+/// Interface for token storage.
+abstract interface class TokenStorage {
+  /// Reads access token from storage.
+  Future<String?> getAccessToken();
+
+  /// Saves access token to storage.
+  Future<void> saveAccessToken(String token);
+
+  /// Clears access token from storage.
+  Future<void> deleteAccessToken();
+}